requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
Automat==20.2.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.5.0
Click==7.0
constantly==15.1.0
cookies==2.2.1
cryptography==2.8
dbsession==0.5.2
distlib==0.3.0
Django==2.1.7
dnspython==1.16.0
falcon==2.0.0
filelock==3.0.12
Flask==0.12.2
Flask-Cors==3.0.3
flask-marshmallow==0.8.0
Flask-RESTful==0.3.6
flask-restful-swagger-2==0.35
Flask-SQLAlchemy==2.3.2
flask-swagger-ui==0.0.3
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.0.2
incremental==17.5.0
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.0
jaraco.text==3.2.0
Jinja2==2.9.6
jsonschema==2.6.0
jwt==0.6.1
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.13.2
more-itertools==8.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.19
ndjson==0.3.1
numpy==1.18.1
olefile==0.46
packaging==20.1
paramiko==2.7.1
Pillow==6.0.0
pluggy==0.13.1
portend==2.6
protobuf==3.11.2
py==1.8.1
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
PyMySQL==0.8.1
PyNaCl==1.3.0
pyparsing==2.4.6
pytest==5.3.5
pytest-dotenv==0.4.0
python-dotenv==0.10.5
python-magic==0.4.15
pytz==2019.3
PyYAML==3.12
redis==3.4.1
requests==2.22.0
safrs==1.0.17
six==1.14.0
SQLAlchemy==1.2.2
sqlalchemy-easy-profile==1.0.3
SQLAlchemy-ImageAttach==1.1.0
sqlalchemy-media==0.17.4
sqlparse==0.3.1
tempora==2.1.0
toml==0.10.0
tox==3.14.5
Twisted==19.10.0
urllib3==1.25.8
validate-email==1.3
virtualenv==20.0.7
Wand==0.5.9
wcwidth==0.1.8
Werkzeug==0.14.1
zc.lockfile==2.0
zipp==3.0.0
zope.interface==4.7.1
